Formatting and Some Functionality
Re-formatted registerform to be more readable (Should rename to signupform).
Also added functionality to the sign up and log in label at the bottom of each window which switches the user to the other.

Todo: Add a style class in another file so that they don't have to be repeated over each form. Also rework how app is being transferred perhaps.